refactoring and extended properties and properties details components

+++ EstateManagement/ClientApp/app/components/properties/components/properties.component.ts
@@ -4,26 +4,36 @@ import { PropertiesService } from '../../properties/services/PropertiesService';
 import { Router } from '@angular/router';
 import { Inject } from '@angular/core';
 import { ProgressSpinnerModule } from 'primeng/progressspinner';
+import { ConfirmationService } from 'primeng/api';
+import { Location } from '@angular/common';
+import { ActivatedRoute } from '@angular/router';
+import { BaseComponent } from '../../../common/base.component';
+import { Message } from 'primeng/components/common/api';
+
 
 
 @Component({
-    templateUrl: './properties.component.html'
+    templateUrl: './properties.component.html',
+    providers: [ConfirmationService],
 })
 
-export class PropertiesComponent implements OnInit {
+export class PropertiesComponent extends BaseComponent implements OnInit {
   
 
     constructor(
+        private confirmationService: ConfirmationService,
         private propertiesService: PropertiesService,
         private router: Router,
-    ) { };
+        private location: Location,
+        private activatedRoute: ActivatedRoute,
+    ) { super(activatedRoute, location) };
 
     
     properties: Array<Property> = new Array<Property>();
     pageTitle: string = "List of available properties";
-    tempInfo: string = "Loading...";
    
     ngOnInit(): void {
+        this.messages = new Array<Message>();
         this.downloadProperties();
 
     }
@@ -32,15 +42,8 @@ export class PropertiesComponent implements OnInit {
 
     downloadProperties(): void {
         this.propertiesService.getAllProperties().subscribe(
-            propertiesFromDb => {
-                if (propertiesFromDb.length == 0) {
-                    this.tempInfo = "Records not found.";
-                }
-                else {
-                    this.properties = propertiesFromDb;
-                }
-            },
-            error => console.log(error)
+            propertiesFromDb => this.properties = propertiesFromDb,
+            errorMessage => this.showMassage(true, 'warn', 'Information', false, errorMessage)
         );
     }
 
@@ -53,13 +56,25 @@ export class PropertiesComponent implements OnInit {
     }
 
     deleteProperty(id: number): void {
-        this.propertiesService.deleteProperty(id).subscribe(
-            onSuccess => {
-                console.log(onSuccess);
-                this.properties.splice(this.properties.findIndex(prop => prop.Id == id), 1);
-                },
-            onError => console.log(onError)
-        );
+        this.confirmationService.confirm({
+            message: 'Are you sure that want to delete this property?',
+            header: 'Confirmation',
+            icon: 'fa fa-question-circle',
+            accept: () => {
+                this.propertiesService.deleteProperty(id).subscribe(
+                    onSuccess => {
+                        this.showMassage(false, 'success', 'Confirmation', true, 'Property has been deleted successfully!')
+                        this.properties.splice(this.properties.findIndex(prop => prop.Id == id), 1);
+                    },
+                    errorMessage => this.showMassage(true, 'warn', 'Information', false, errorMessage)
+                );
+            },
+            reject: () => {
+                //Jeżeli zrezygnuje to co ma się wykonać ?
+            }
+        })
+
+        
         
     }
